# main-api/swagger.yml
openapi: 3.0.0
paths:
  /image/{filename}:
    get:
      operationId: AppController_getImage
      parameters:
        - name: filename
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageEntity'
      tags: &ref_0
        - app
  /latest-results:
    get:
      operationId: AppController_getLatestResults
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AiAnalysisLogEntity'
      tags: *ref_0
  /save-analysis-image:
    post:
      operationId: AppController_saveAndAnalysisImage
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveAndAnalysisImageRequest'
      responses:
        '200':
          description: ''
      tags: *ref_0
info:
  title: backend API
  description: backend API description
  version: '1.0'
  contact: {}
tags:
  - name: backend
    description: ''
servers: []
components:
  securitySchemes:
    jwt:
      scheme: bearer
      bearerFormat: JWT
      type: http
      name: jwt
      description: Enter JWT token
      in: header
  schemas:
    AiAnalysisLogEntity:
      type: object
      properties:
        class:
          type: string
          nullable: true
        confidence:
          type: number
          nullable: true
        id:
          type: number
        imagePath:
          type: string
          nullable: true
        message:
          type: string
          nullable: true
        requestTimestamp:
          format: date-time
          type: string
          nullable: true
        responseTimestamp:
          format: date-time
          type: string
          nullable: true
        success:
          type: string
          nullable: true
      required:
        - class
        - confidence
        - id
        - imagePath
        - message
        - requestTimestamp
        - responseTimestamp
        - success
    ImageEntity:
      type: object
      properties:
        data:
          type: string
          nullable: true
        filePath:
          type: string
          nullable: true
      required:
        - data
        - filePath
    SaveAndAnalysisImageRequest:
      type: object
      properties:
        imgData:
          type: string
      required:
        - imgData